pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oci-vision-ai"
version = "0.1.0"
description = "The definitive OCI Vision AI showcase — CLI, web dashboard, and notebooks"
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.10"
dependencies = [
    "oci>=2.110.0",
    "pydantic>=2.0",
    "typer>=0.9.0",
    "rich>=13.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "Pillow>=10.0",
]

[project.optional-dependencies]
notebooks = ["matplotlib>=3.8", "jupyter>=1.0", "pandas>=2.0"]
dev = ["pytest>=7.4", "pytest-asyncio>=0.23", "pytest-cov>=4.1", "httpx>=0.25"]
oracle = ["oracledb>=2.4.0"]
all = ["oci-vision-ai[notebooks,dev,oracle]"]

[project.scripts]
oci-vision = "oci_vision.cli.app:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "live: requires explicit OCI opt-in via OCI_VISION_RUN_LIVE=1",
    "oracle: requires Oracle Database 26ai integration env via OCI_VISION_RUN_ORACLE=1",
]
